requirements.txt
django==4.1.13
dotenv==0.9.9
requests==2.32.0
boto3==1.38.2
django-storages==1.14
qtfaststart==1.8
gunicorn==23.0.0
psycopg2-binary==2.9.10
whitenoise==6.9.0
dj-database-url==0.5.0
PyJWT==2.8.0